\name{SimStudy}

\alias{SimStudy}

\title{
A simulation study on penalized B-splines' MSE performance
}

\description{
Compare MSE performance of O-spline ("os"), standard P-spline ("sps") and general P-spline ("gps") via a simulation study with random curves generated by \code{\link{RandomSpl}}.
}

\usage{
SimStudy(n, degree, periodic = FALSE, n2s.ratio = 0.2, n.sim = 100, nc = 1)
}

\arguments{
\item{n}{number of unevenly spaced \eqn{x}-values. This argument is passed to \code{\link{RandomSpl}}.}
\item{degree}{polynomial degree of the spline.}
\item{periodic}{TRUE to generate a periodic spline; FALSE to generate an ordinary spline. This argument is passed to \code{\link{RandomSpl}}.}
\item{n2s.ratio}{a noise-to-signal ratio that is used when simulating noisy \eqn{y}-values.}
\item{n.sim}{number of simulations.}
\item{nc}{number of CPU cores to use for parallel simulations.}
}

\details{
See Section 3 of the reference.
}

\value{
A matrix of MSE values. It has \code{n.sim} rows and \code{(3 * degree)} named columns. A name is like "spl.m", where "spl" is an abbreviation ("os", "sps" or "gps") and "m" is a penalty order from 1 to \code{degree}.
}

\author{
Zheyuan Li \email{zheyuan.li@bath.edu}
}

\references{
Li, Z. and Cao, J. (2022). General P-Splines for Non-Uniform B-Splines.
}

\examples{
\dontrun{

require(gps.mgcv)

SimStudy(n = 400, degree = 3, n2s.ratio = 0.2, n.sim = 5)
}
}
